workflows

+++ build-wheels.sh
@@ -7,6 +7,10 @@ export PATH="$HOME/.cargo/bin:$PATH"
 
 # Compile wheels
 for PYBIN in /opt/python/cp{310,311,312,313}*/bin; do
+    if [[ $PYBIN == *"313t"* ]]; then
+      echo "Skipping experimental Python 3.13t"
+      continue
+    fi
     rm -rf /io/build/
     "${PYBIN}/pip" install -U setuptools setuptools-rust wheel
     "${PYBIN}/pip" wheel /io/ -w /io/dist/ --no-deps
